chore: typescripify `canary-release.js` (#634)
