+++ patches/@changesets+apply-release-plan+5.0.1.patch
@@ -1,5 +1,5 @@
 diff --git a/node_modules/@changesets/apply-release-plan/dist/apply-release-plan.cjs.dev.js b/node_modules/@changesets/apply-release-plan/dist/apply-release-plan.cjs.dev.js
-index 8b33aee..09ebd85 100644
+index 5435d7a..36faeed 100644
 --- a/node_modules/@changesets/apply-release-plan/dist/apply-release-plan.cjs.dev.js
 +++ b/node_modules/@changesets/apply-release-plan/dist/apply-release-plan.cjs.dev.js
 @@ -96,7 +96,8 @@ function shouldUpdateDependencyBasedOnConfig(release, {
@@ -29,7 +29,7 @@ index 8b33aee..09ebd85 100644
            if (usesWorkspaceRange) newNewRange = `workspace:${newNewRange}`;
            deps[name] = newNewRange;
          }
-@@ -271,7 +273,7 @@ async function getCommitsThatAddChangesets(changesetIds, cwd) {
+@@ -273,7 +275,7 @@ async function getCommitsThatAddChangesets(changesetIds, cwd) {
    return commits;
  }
  
@@ -38,7 +38,7 @@ index 8b33aee..09ebd85 100644
    let cwd = packages.root.dir;
    let touchedFiles = [];
    const packagesByName = new Map(packages.packages.map(x => [x.packageJson.name, x]));
-@@ -309,6 +311,7 @@ async function applyReleasePlan(releasePlan, packages, config$1 = config.default
+@@ -311,6 +313,7 @@ async function applyReleasePlan(releasePlan, packages, config$1 = config.default
    let finalisedRelease = releaseWithChangelogs.map(release => {
      return versionPackage(release, versionsToUpdate, {
        updateInternalDependencies: config$1.updateInternalDependencies,
@@ -47,7 +47,7 @@ index 8b33aee..09ebd85 100644
        bumpVersionsWithWorkspaceProtocolOnly: config$1.bumpVersionsWithWorkspaceProtocolOnly
      });
 diff --git a/node_modules/@changesets/apply-release-plan/dist/apply-release-plan.cjs.prod.js b/node_modules/@changesets/apply-release-plan/dist/apply-release-plan.cjs.prod.js
-index 9e9d03c..fff3e27 100644
+index f83761f..bc94382 100644
 --- a/node_modules/@changesets/apply-release-plan/dist/apply-release-plan.cjs.prod.js
 +++ b/node_modules/@changesets/apply-release-plan/dist/apply-release-plan.cjs.prod.js
 @@ -64,7 +64,7 @@ function shouldUpdateDependencyBasedOnConfig(release, {depVersionRange: depVersi
@@ -68,7 +68,7 @@ index 9e9d03c..fff3e27 100644
          usesWorkspaceRange && (newNewRange = "workspace:" + newNewRange), deps[name] = newNewRange;
        }
      }
-@@ -160,7 +160,7 @@ async function getCommitsThatAddChangesets(changesetIds, cwd) {
+@@ -161,7 +161,7 @@ async function getCommitsThatAddChangesets(changesetIds, cwd) {
    })), commits;
  }
  
@@ -77,7 +77,7 @@ index 9e9d03c..fff3e27 100644
    let cwd = packages.root.dir, touchedFiles = [];
    const packagesByName = new Map(packages.packages.map((x => [ x.packageJson.name, x ])));
    let {releases: releases, changesets: changesets} = releasePlan;
-@@ -177,6 +177,7 @@ async function applyReleasePlan(releasePlan, packages, config$1 = config.default
+@@ -178,6 +178,7 @@ async function applyReleasePlan(releasePlan, packages, config$1 = config.default
      type: type
    }))), finalisedRelease = releaseWithChangelogs.map((release => versionPackage(release, versionsToUpdate, {
      updateInternalDependencies: config$1.updateInternalDependencies,
@@ -85,3 +85,12 @@ index 9e9d03c..fff3e27 100644
      onlyUpdatePeerDependentsWhenOutOfRange: config$1.___experimentalUnsafeOptions_WILL_CHANGE_IN_PATCH.onlyUpdatePeerDependentsWhenOutOfRange,
      bumpVersionsWithWorkspaceProtocolOnly: config$1.bumpVersionsWithWorkspaceProtocolOnly
    }))), prettierConfig = await prettier__default.default.resolveConfig(cwd);
+diff --git a/node_modules/@changesets/apply-release-plan/dist/declarations/src/index.d.ts b/node_modules/@changesets/apply-release-plan/dist/declarations/src/index.d.ts
+index db7ebee..7602b71 100644
+--- a/node_modules/@changesets/apply-release-plan/dist/declarations/src/index.d.ts
++++ b/node_modules/@changesets/apply-release-plan/dist/declarations/src/index.d.ts
+@@ -1,3 +1,3 @@
+ import { ReleasePlan, Config } from "@changesets/types";
+ import { Packages } from "@manypkg/get-packages";
+-export default function applyReleasePlan(releasePlan: ReleasePlan, packages: Packages, config?: Config, snapshot?: string | boolean): Promise<string[]>;
++export default function applyReleasePlan(releasePlan: ReleasePlan, packages: Packages, config?: Config, snapshot?: string | boolean, ignoreRange: boolean): Promise<string[]>;
